Exception handler

+++ src/main/java/com/jmp/dscommerce/services/ProductService.java
@@ -2,18 +2,23 @@ package com.jmp.dscommerce.services;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.jmp.dscommerce.dtos.ProductDto;
 import com.jmp.dscommerce.entities.Product;
 import com.jmp.dscommerce.repositories.ProductRepository;
+import com.jmp.dscommerce.services.exceptions.DatabaseException;
+import com.jmp.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -22,18 +27,29 @@ public class ProductService {
 	private ProductRepository repository;
 
 
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)	
 	public void delete(Long id) {
-		repository.deleteById(id);		
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso nao encontrado.");
+		}
+		try {
+			repository.deleteById(id);			
+		}catch(DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial.");
+		}		
 	}
 	
 	@Transactional
 	public ProductDto update(Long id, ProductDto dto) {
+		try {
 		Product entity = repository.getReferenceById(id);
 		copyDtoToEntity(dto, entity);		
 		entity = repository.save(entity);
 		
-		return 	new ProductDto(entity);		
+		return 	new ProductDto(entity);
+		}catch(EntityNotFoundException e){
+			throw new ResourceNotFoundException("Recurso nao encontrado");
+		}
 	}
 	
 	
@@ -61,8 +77,8 @@ public class ProductService {
 	
 	@Transactional(readOnly=true)
 	public ProductDto findById(Long id) {
-		Optional<Product> result = repository.findById(id);
-		Product product = result.get();
+		Product product = repository.findById(id).orElseThrow(
+				()-> new ResourceNotFoundException("Recurso nao encontrado"));
 		
 		ProductDto dto = new ProductDto(product);
